migrated to nginx charmlib
