+++ pyproject.toml
@@ -15,6 +15,7 @@ requires-python = ">=3.10"
 # NOTE: always replicate any changes in dependencies to Coordinator and Worker tester charms as they should always match this
 dependencies = [
     "charmlibs-interfaces-tls-certificates~=1.0",
+    "charmlibs-nginx-k8s>=0.1.0",
     "crossplane",
     "ops[tracing]",
     "pydantic",
@@ -22,6 +23,7 @@ dependencies = [
     "charmed-service-mesh-helpers>=0.2.0",
     "lightkube-extensions",
     "cosl>=1.3.1",
+    "charmlibs-nginx-k8s>=0.1.0",
 ]
 classifiers = [
     "Programming Language :: Python :: 3.10",
